requirements.txt
# Core Dependencies
numpy>=1.21.0
Pillow>=9.0.0
PyQt6>=6.4.0
opencv-python>=4.5.0
scipy>=1.7.0
numba>=0.54.0
pyopengl>=3.1.0
pycuda>=2021.1
pyopencl>=2021.1
